Move IO to HTTP implementation
Remove InputStream from APIs.

When serializing a Resource to JSON, having InputStream as a result
makes it unclear when exactly the serialization happens.
Serialization is also not real IO, just an in-memory operation.

When reading the HTTP response to a Resource, we really want to do the
slow IO in the HTTP thread.  This makes it HTTP implementation
responsibility to close the socket.  Also, this allows JSON
deserialization be decoupled from the IO.