Check the property type when building a resource
If the property doesn't exist, set a dynamic property.
If the resource is not a base class, trace log the action.

If the property is assignable, then do just that.

Otherwise, try to "groovy cast" the value.
Log an error (instead of throwing an exception) on failure to cast.

Also, fix the found typos and type mismatches.